feat(skill) : added delete skill and skill category feature

+++ src/main/java/com/mohit/portfolio_backend/controller/Admin/AdminSkillCategoryController.java
@@ -2,7 +2,9 @@ package com.mohit.portfolio_backend.controller.Admin;
 
 import com.mohit.portfolio_backend.entity.SkillCategory;
 import com.mohit.portfolio_backend.repository.SkillCategoryRepository;
+import com.mohit.portfolio_backend.repository.SkillRepository;
 import lombok.RequiredArgsConstructor;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -13,6 +15,7 @@ import java.util.List;
 public class AdminSkillCategoryController {
 
     private final SkillCategoryRepository repository;
+    private final SkillRepository skillRepository;
 
     @PostMapping
     public SkillCategory create(@RequestBody SkillCategory category) {
@@ -25,7 +28,23 @@ public class AdminSkillCategoryController {
     }
 
     @DeleteMapping("/{id}")
-    public void delete(@PathVariable Long id) {
+    public ResponseEntity<?> delete(@PathVariable Long id) {
+
+        // Check if category exists
+        if (!repository.existsById(id)) {
+            return ResponseEntity.notFound().build();
+        }
+
+        // Check if any skill is using this category
+        if (skillRepository.existsByCategoryId(id)) {
+            return ResponseEntity
+                    .badRequest()
+                    .body("Cannot delete category. Skills are still assigned.");
+        }
+
+        // Safe delete
         repository.deleteById(id);
+
+        return ResponseEntity.noContent().build();
     }
 }
